Reorganiza e melhora renderização
