+++ _scss/styles.scss
@@ -139,8 +139,3 @@ $legend-font-weight: 500;
 #home-hero {
   height: 50vh;
 }
-
-.triptych {
-  margin-bottom: 1.5rem;
-  text-align: center;
-}
