+++ package.json
@@ -27,8 +27,8 @@
 		"11ty": "npx @11ty/eleventy",
 		"dev": "npx @11ty/eleventy --serve",
 		"css": "npm-run-all css-compile css-prefix",
-		"css-compile": "sass --style compressed --source-map --embed-sources --no-error-css --load-path=node_modules scss/:assets/css/",
-		"css-lint": "stylelint scss/",
+		"css-compile": "sass --style compressed --source-map --embed-sources --no-error-css --load-path=node_modules _scss/:assets/css/",
+		"css-lint": "stylelint _scss/",
 		"css-prefix": "postcss --replace assets/css/styles.css --use autoprefixer --map",
 		"watch": "nodemon -e html,scss -x \"npm run css\"",
 		"start": "npm-run-all --parallel watch dev",
