+++ src/_data/ui.json
@@ -1,6 +1,12 @@
 {
+  "architectural_documents": "documentação arquitetônica",
+  "bibliography": "referências bibliográficas",
+  "iconography": "iconografia",
+  "primary_sources": "fontes primárias",
   "read_more": "Leia mais",
   "search_verb": "Pesquisar",
-  "br": "Brasil",
-  "pt": "Portugal"
+  "BR": "Brasil",
+  "CV": "Cabo Verde",
+  "IN": "Índia",
+  "PT": "Portugal"
 }
