+++ src/casa/br-df-planaltina-mhap.md
@@ -1,3 +1,9 @@
 ---
 title: "Museu Histórico e Artístico de Planaltina"
+author: Pedro P. Palazzo
+date: 2024-02-08
+location:
+  city: "Planaltina"
+  state: "Distrito Federal"
+  country: BR
 ---
