+++ src/index.html
@@ -2,6 +2,9 @@
 title: "Página inicial"
 layout: "layouts/home.njk"
 isHome: true
+header:
+  overlay_image: "https://i.pinimg.com/originals/f7/76/49/f7764902ee90d4fc48b6f795307bc366.jpg"
+  caption: "Abigail de Andrade, Estrada do Mundo Novo, 1888"
 triptych:
 - image_path: https://placehold.co/140x140/png
   title     : "Planalto central"
@@ -23,10 +26,10 @@ triptych:
 <section id="colecoes-especiais">
   <h2 class="text-center">Coleções especiais</h2>
 
-  <div class="container triptych">
-    <div class="row">
+  <div class="container">
+    <div class="row my-2 text-center">
       {% for t in triptych %}
-      <div class="col-lg-4">
+      <div class="col-md-6 col-lg-4">
         {% if t.image_path %}
           <img class="rounded-circle" src="{{ t.image_path }}" width="140" height="140" alt="{{ t.title }}" />
         {% endif %}
@@ -48,6 +51,10 @@ triptych:
 <section id="inclusoes-recentes">
   <h2 class="text-center">Inclusões recentes</h2>
 
-  {% assign album = collections.casa %}
-  {% include "partials/album.njk" %}
+  <div class="container-fluid">
+    <div class="row g-3 mx-3">
+    {% assign album = collections.casa | reverse %}
+    {% include "partials/album.njk" %}
+    </div><!--/.row-->
+  </div><!--/.container-fluid-->
 </section>
