+++ src/pages/casa.md
@@ -1,4 +1,5 @@
 ---
 title: "Lista de casas"
 layout: "layouts/base.njk"
+permalink: /casa/
 ---
